requirements.txt
daemonize
urllib3
python-dotenv
